Rename lint task to linthtml
